scripts/requirements.txt
# SaveTheWorld Python dependencies
anthropic>=0.39.0
pyyaml>=6.0
